SonarQube batch 7: JS/C++/Python quality fixes (~65 issues)
JavaScript rules fixed:
- S7773 (8): isNaN -> Number.isNaN across 4 files
- S7761 (9): hasAttribute/setAttribute -> .dataset in SequenceController.js
- S7781 (8): .replace(/x/g) -> .replaceAll() in filesystem.html
- S7735 (2): Negated conditions -> positive/nullish coalescing
- S3358 (3): Nested ternaries -> if/else in filesystem.html + ToolsController.js
- S7762 (1): parentNode.removeChild -> .remove() in filesystem.html
- S7765 (1): .indexOf === -1 -> .includes() in index.html
- S7764 (1): window -> globalThis in index.html
- S2486 (7): Empty catch blocks -> handled/documented
- S7718 (1): Catch param e2 -> error_ in i18n.js
- S3735 (2): void operator -> direct expression in utils.js/StatsController.js
- S125 (1): Commented-out code -> descriptive text in utils.js
- S1481+S1854 (2): Unused var + dead store in setup.html

C++ rules fixed:
- S5817 (8): const methods in ContactSensors (4) + PursuitController (4)
- S995 (4): const SequenceLine* params in SequenceExecutor start methods
- S5276 (2): Explicit casts in ChaosController + SequenceTableManager
- S125 (1): Reformulated comment in APIRoutes.cpp

Python rules fixed:
- S1481 (2): Removed unused vars (lines, rel) in find_dead_code.py
- S1764 (1): Fixed name == name tautology in find_dead_code.py
- S5754 (1): except: -> except Exception: in find_dead_code.py
- S3358 (1): Nested ternary -> if/elif/else in analyze_sonarqube.py

Build: SUCCESS | Tests: 180/180 PASSED

+++ src/movement/ChaosController.cpp
@@ -407,7 +407,7 @@ void ChaosController::handleMultiPhase(const ChaosBaseConfig& cfg, const ChaosMu
     float maxPossibleAmplitude = calculateMaxAmplitude(effectiveMinLimit, effectiveMaxLimit);
     chaosState.waveAmplitude = maxPossibleAmplitude * (cfg.amplitudeJumpMin + (cfg.amplitudeJumpMax - cfg.amplitudeJumpMin) * craziness);
 
-    int forwardChance = dir_cfg.forwardChanceMin - static_cast<int>((dir_cfg.forwardChanceMin - dir_cfg.forwardChanceMax) * craziness);
+    int forwardChance = dir_cfg.forwardChanceMin - static_cast<int>(static_cast<float>(dir_cfg.forwardChanceMin - dir_cfg.forwardChanceMax) * craziness);
     chaosState.movingForward = random(100) < forwardChance;
 
     setDirectionalTarget(chaosState.waveAmplitude, effectiveMinLimit, effectiveMaxLimit);
